pyproject.toml
[tool.poetry]
name = "stool"
version = "0.1.0"
description = ""
authors = ["Stanford <bigfatsea@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
python-dateutil = "^2.9.0.post0"
colorlog = "^6.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
